+++ dosma/core/__init__.py
@@ -1,0 +1,11 @@
+from dosma.core import (  # noqa: F401
+    med_volume,
+    orientation,
+)
+
+from dosma.core.med_volume import *  # noqa
+from dosma.core.orientation import *  # noqa
+
+__all__ = []
+__all__.extend(med_volume.__all__)
+__all__.extend(orientation.__all__)
