+++ dosma/core/io/__init__.py
@@ -1,0 +1,16 @@
+from dosma.core.io import (  # noqa: F401
+    dicom_io,
+    format_io,
+    format_io_utils,
+    nifti_io,
+)
+from dosma.core.io.dicom_io import *  # noqa
+from dosma.core.io.format_io import *  # noqa
+from dosma.core.io.format_io_utils import *  # noqa
+from dosma.core.io.nifti_io import *  # noqa
+
+__all__ = []
+__all__.extend(dicom_io.__all__)
+__all__.extend(format_io.__all__)
+__all__.extend(format_io_utils.__all__)
+__all__.extend(nifti_io.__all__)
