--- dosma/data_io/__init__.py
@@ -1,20 +1,0 @@
-from dosma.data_io import (  # noqa: F401
-    dicom_io,
-    format_io,
-    format_io_utils,
-    med_volume,
-    nifti_io,
-    orientation,
-)
-from dosma.data_io.dicom_io import *  # noqa
-from dosma.data_io.format_io import *  # noqa
-from dosma.data_io.format_io_utils import *  # noqa
-from dosma.data_io.nifti_io import *  # noqa
-from dosma.data_io.med_volume import *  # noqa
-
-__all__ = []
-__all__.extend(dicom_io.__all__)
-__all__.extend(format_io.__all__)
-__all__.extend(format_io_utils.__all__)
-__all__.extend(med_volume.__all__)
-__all__.extend(nifti_io.__all__)
