+++ dosma/scan_sequences/qdess.py
@@ -6,7 +6,7 @@ from typing import Sequence, Tuple
 
 from pydicom.tag import Tag
 
-from dosma.data_io.med_volume import MedicalVolume
+from dosma.core.med_volume import MedicalVolume
 from dosma.models.seg_model import SegModel
 from dosma.quant_vals import T2
 from dosma.scan_sequences.scans import ScanSequence
