+++ dosma/utils/fits.py
@@ -13,7 +13,7 @@ from tqdm.auto import tqdm
 from tqdm.contrib.concurrent import process_map
 
 from dosma import defaults
-from dosma.data_io.med_volume import MedicalVolume
+from dosma.core.med_volume import MedicalVolume
 from dosma.defaults import preferences
 from dosma.utils import env
 from dosma.utils.device import cpu_device, get_array_module, get_device
@@ -33,8 +33,7 @@ __all__ = [
 
 
 class _Fit(ABC):
-    """Abstract class for fitting quantitative values.
-    """
+    """Abstract class for fitting quantitative values."""
 
     @abstractmethod
     def fit(self) -> Tuple[MedicalVolume, MedicalVolume]:
