+++ tests/utils/test_fitting.py
@@ -2,7 +2,7 @@ import unittest
 
 import numpy as np
 
-from dosma.data_io.med_volume import MedicalVolume
+from dosma.core.med_volume import MedicalVolume
 from dosma.utils.fits import (
     CurveFitter,
     MonoExponentialFit,
